Update(Refactoring): Refactoring the Enemy classes
EnemyDisplay: The anonymous method in the "Initialize" method was replaced with the "HideHealthSlider" method.

EnemyKeeper: Unnecessary fields were removed. The "GetEnemy" method was simplified cause its code duplicates the code of the "CanGetEnemy" method to check for the enemies. Were added condition to check if there are enemies in the "RemoveKilledEnemy" method before it removes an enemy from list.

EnemySaver: The "EnemySaver" class is no longer inherited from MonoBehaviour anymore because this isn`t necessary. Unused fields were removed. The "Construct" method was replaced by the constructor of the EnemySaver class. the "x" parameter in the 91 line was replaced with "serializeEnemy" for greater clarity.